proc: cast p_align to size_t

+++ proc/process.c
@@ -378,7 +378,7 @@ static int process_validateElf32(void *iehdr, size_t size)
 		}
 
 		offs = (off_t)(Elf32_Off)(phdr[i].p_offset & ~(phdr[i].p_align - 1U));
-		misalign = ((size_t)phdr[i].p_offset & ((size_t)phdr[i].p_align - 1U));
+		misalign = ((size_t)phdr[i].p_offset & (size_t)(phdr[i].p_align - 1U));
 		filesz = (phdr[i].p_filesz != 0U) ? (phdr[i].p_filesz + misalign) : 0;
 		memsz = phdr[i].p_memsz + misalign;
 		if ((offs >= (off_t)size) || (memsz < filesz)) {
@@ -454,7 +454,7 @@ static int process_validateElf64(void *iehdr, size_t size)
 		}
 
 		offs = (off_t)(Elf64_Off)(phdr[i].p_offset & ~(phdr[i].p_align - 1U));
-		misalign = (size_t)phdr[i].p_offset & (phdr[i].p_align - 1U);
+		misalign = (size_t)phdr[i].p_offset & (size_t)(phdr[i].p_align - 1U);
 		filesz = (phdr[i].p_filesz != 0U) ? (size_t)(phdr[i].p_filesz + misalign) : 0UL;
 		memsz = (size_t)phdr[i].p_memsz + misalign;
 		if ((offs >= (off_t)size) || (memsz < filesz)) {
@@ -545,7 +545,7 @@ static int process_load32(vm_map_t *map, vm_object_t *o, off_t base, void *iehdr
 
 		vaddr = (void *)(((ptr_t)phdr->p_vaddr & ~(phdr->p_align - 1U)));
 		offs = (off_t)(unsigned int)(phdr->p_offset & ~(phdr->p_align - 1U));
-		misalign = (size_t)phdr->p_offset & ((size_t)phdr->p_align - 1U);
+		misalign = (size_t)phdr->p_offset & (size_t)(phdr->p_align - 1U);
 		filesz = (phdr->p_filesz != 0U) ? (size_t)(u32)(phdr->p_filesz + misalign) : 0UL;
 		memsz = (size_t)(u32)(phdr->p_memsz + misalign);
